l6as: implement assembler output
- Add binary and listing outputs

+++ l6as/src/assembler/size.rs
@@ -1,7 +1,7 @@
 use super::statements::{BranchLocation, BranchOnIndicatorsOpCode, Statement};
 
 /// Computes the size of a statement in memory (in words)
-pub fn statement_size(statement: &Statement, _cur_addr: u128) -> u128 {
+pub fn statement_size(statement: &Statement, _cur_addr: u64) -> u64 {
     // Compute different size depending on the kind of statement
     match statement {
         Statement::Org(_) => 0,
@@ -15,12 +15,12 @@ pub fn statement_size(statement: &Statement, _cur_addr: u128) -> u128 {
 pub fn branch_on_indicators_inst_size(
     _op: &BranchOnIndicatorsOpCode,
     branchloc: &BranchLocation,
-) -> u128 {
+) -> u64 {
     1 + branchloc_extra_words(branchloc)
 }
 
 /// Computes amount of extra words used by a Branch location
-pub fn branchloc_extra_words(branchloc: &BranchLocation) -> u128 {
+pub fn branchloc_extra_words(branchloc: &BranchLocation) -> u64 {
     match branchloc {
         BranchLocation::Absolute(_) => 1,
         BranchLocation::LongDisplacement(_) => 1,
